backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
aiogram>=3.4
python-dotenv>=1.0
openai>=1.0
